creating PasswordPolicyAlert component
