+++ src/main/resources/i18n/messages.properties
@@ -529,8 +529,6 @@ UserSecurityPage.alert.title=Missing Email Configuration
 UserSecurityPage.alert.description=Password reset is disabled due to this instance of IRIDA not having a configured mail server.
 
 UserChangePasswordForm.title=Change Password
-UserChangePasswordForm.alert.title=Password Policy
-UserChangePasswordForm.alert.description=Please choose a strong password that meets the following criteria:
 UserChangePasswordForm.form.label.oldPassword=Old Password
 UserChangePasswordForm.form.label.newPassword=New Password
 UserChangePasswordForm.form.button.submit=Submit
@@ -2501,8 +2499,6 @@ CreateNewUser.emailConfigured.alert.title=Missing Email Configuration
 CreateNewUser.emailConfigured.alert.description=User activation emails are disabled due to this instance of IRIDA not having a configured mail server. Please enter a password for the new user account.
 CreateNewUser.mailFailure.alert.title=Email Failure
 CreateNewUser.mailFailure.alert.description=IRIDA was unable to send an activation e-mail. Please enter a password for the new user account.
-CreateNewUser.changePassword.alert.title=Password Policy
-CreateNewUser.changePassword.alert.description=Please choose a strong password that meets the following criteria:
 CreateNewUser.notification.success=New user created successfully.
 CreateNewUser.notification.error=An error has occurred while creating the new user. Please check all the fields.
 
@@ -2860,19 +2856,18 @@ ActivateAccount.alert.description=The Activation ID can be found in the welcome 
 # PasswordReset Component                                                                    #
 # ========================================================================================== #
 PasswordReset.title=Reset Password
-PasswordReset.alert.title=Password Policy
-PasswordReset.alert.description=Please choose a strong password that meets the following criteria:
-PasswordReset.alert.recommendation1=Passwords for admins are recommended to be atleast 11 characters
-PasswordReset.alert.recommendation2=Passwords should not form any words or contain any personal information (recommended)
-PasswordReset.passwordIsRequired=Password is required
 PasswordReset.input.placeholder=New Password
 PasswordReset.button.setPassword=Set Password
-PasswordReset.input.passwordNotMatch=Password does not match the password policy
-PasswordReset.input.minLength=Password must be a minimum of 8 characters
 PasswordReset.followingErrorsOccurred=The following errors occurred:
 PasswordReset.alert.success=Password successfully updated.  You may use your new credentials to log in to IRIDA.
 PasswordReset.alert.success.link=Return to IRIDA
 PasswordReset.password.expired=Your password has expired.  Please create a new password.
+
+# ========================================================================================== #
+# PasswordPolicyAlert Component                                                                    #
+# ========================================================================================== #
+PasswordPolicyAlert.title=Password Policy
+PasswordPolicyAlert.description=Please choose a strong password that meets the following criteria:
 
 # ========================================================================================== #
 # Utilities                                                                                   #
